+++ src/main/clojure/com/github/clojure_repl/intellij/keyboard_manager.clj
@@ -1,10 +1,7 @@
 (ns com.github.clojure-repl.intellij.keyboard-manager
   (:require
-   [com.github.ericdallo.clj4intellij.app-manager :as app-manager]
    [com.rpl.proxy-plus :refer [proxy+]])
   (:import
-   [com.intellij.find SearchReplaceComponent]
-   [com.intellij.ide IdeEventQueue]
    [com.intellij.openapi.editor Editor]
    [java.awt KeyEventDispatcher KeyboardFocusManager]
    [java.awt.event KeyEvent]))
@@ -18,11 +15,7 @@
 
                      (dispatchKeyEvent [_ ^KeyEvent event]
                        (if (and (= KeyEvent/KEY_PRESSED (.getID event))
-                                (or (= (.getContentComponent editor) (.getComponent event))
-                                    ;; Intellij shows the searchReplace ignoring this listener
-                                    ;; so we send the event even if this search window compoennt
-                                    ;; was opened.
-                                    (= SearchReplaceComponent (type (.getComponent event)))))
+                                (= (.getContentComponent editor) (.getComponent event)))
                          (boolean (on-key-pressed event))
                          false)))]
     (swap! key-listeners* assoc editor dispatcher)
@@ -35,14 +28,3 @@
     (when (= editor-to-unregister editor)
       (.removeKeyEventDispatcher (KeyboardFocusManager/getCurrentKeyboardFocusManager) dispatcher)
       (swap! key-listeners* dissoc editor))))
-
-(defn send-key-pressed! [^Editor editor ^Integer key-code]
-  (app-manager/invoke-later!
-   {:invoke-fn (fn []
-                 (.dispatchEvent (IdeEventQueue/getInstance)
-                                 (KeyEvent. (.getContentComponent editor)
-                                            KeyEvent/KEY_PRESSED
-                                            (System/currentTimeMillis)
-                                            0
-                                            key-code
-                                            (char key-code))))}))