+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
  - Fix `eval defun at cursor` action error. #121
+ - Create view error on test error. #128
 
 ## 2.1.0